# openapi.json
{
  "components": {
    "schemas": {
...
      "APIKeyCreateRequest": {
        "description": "Create API key request.",
        "properties": {
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "rate_limit_rpm": {
            "default": 60,
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Rate Limit Rpm",
            "type": "integer"
          },
          "scopes": {
            "description": "Permission scopes: inference, fine_tune, datasets, models_read, usage_read",
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          }
        },
        "required": [
          "name"
        ],
        "title": "APIKeyCreateRequest",
        "type": "object"
      },
      "APIKeyCreateResponse": {
        "description": "Response after creating an API key \u2014 includes the plaintext key (shown once).",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "key": {
            "description": "The full API key (shown once, never again)",
            "title": "Key",
            "type": "string"
          },
          "key_prefix": {
            "title": "Key Prefix",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "key",
          "key_prefix",
          "id",
          "name"
        ],
        "title": "APIKeyCreateResponse",
        "type": "object"
      },
      "APIKeyResponse": {
        "description": "API key response (excludes key_hash for security).",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "description": "API key document ID",
            "title": "Id",
            "type": "string"
          },
          "key_prefix": {
            "description": "First 8 chars for identification: shai_xxxx",
            "title": "Key Prefix",
            "type": "string"
          },
          "last_used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "name": {
            "description": "User-friendly label",
            "title": "Name",
            "type": "string"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "rate_limit_rpm": {
            "title": "Rate Limit Rpm",
            "type": "integer"
          },
          "scopes": {
            "description": "Permission scopes",
            "items": {
              "type": "string"
            },
            "title": "Scopes",
            "type": "array"
          },
          "status": {
            "description": "active, revoked, expired",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "key_prefix",
          "name",
          "organization_id",
          "scopes",
          "rate_limit_rpm",
          "status",
          "created_at"
        ],
        "title": "APIKeyResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Standard API error response (RFC 7807-inspired).",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Human-readable error description",
            "title": "Detail"
          },
          "error": {
            "description": "Machine-readable error code",
            "title": "Error",
            "type": "string"
          },
          "status_code": {
            "description": "HTTP status code",
            "title": "Status Code",
            "type": "integer"
          }
        },
        "required": [
          "error",
          "status_code"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "email": "api@shamwari.ai",
      "name": "Nyuchi Web Services",
      "url": "https://shamwari.ai/"
    },
    "description": "Africa's open source AI platform \u2014 API backend.\n\nShamwari AI provides AI inference, model management, and developer APIs purpose-built for African languages and contexts. Built by Nyuchi Africa.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "title": "Shamwari AI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/api-keys": {
      "get": {
        "description": "Returns all API keys for the authenticated user's organization. Key hashes are never exposed \u2014 only the prefix for identification.",
        "operationId": "list_api_keys_v1_api_keys_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/APIKeyResponse"
                  },
                  "title": "Response List Api Keys V1 Api Keys Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List API keys",
        "tags": [
          "api-keys"
        ]
      },
      "post": {
        "description": "Create a new API key. The full key is returned exactly once in the response \u2014 it cannot be retrieved again. Store it securely.",
        "operationId": "create_api_key_v1_api_keys_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIKeyCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid parameters"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create API key",
        "tags": [
          "api-keys"
        ]
      }
    },
    "/v1/api-keys/{key_id}": {
      "delete": {
        "operationId": "revoke_api_key_v1_api_keys__key_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "title": "Key Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Authentication required"
          },
          "404": {
            "description": "Key not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke API key",
        "tags": [
          "api-keys"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Production",
      "url": "https://api.shamwari.ai"
    },
    {
      "description": "Local development",
      "url": "http://localhost:8000"
    }
  ],
...
}